[python] hostruntime: split tensor_class into coherence, buffer and tensor
tensor_class.py had grown to four concerns in one file: detecting the cache
line, tracking which agent owns which byte range, owning an allocation and
reconciling it, and the tensor surface over all of it. Split along those seams.

  coherence.py     granule detection and _CoherenceMap
  buffer.py        NpuBuffer, CPUBuffer, _TensorBackedBuffer
  torch_interop.py the zero-copy numpy/torch dtype bridge
  tensor_class.py  _WriteBorrow, NpuTensor, CPUOnlyTensor

The dependency runs one way, coherence to buffer to tensor, with no cycle. Every
name that anything imports from the old module is re-exported from it, so no
import in or out of tree changes. Three private constants that moved are not
re-exported (_SYSFS_CACHE_DIR, _UINT_VIEW_DTYPE, _ML_DTYPE_TO_TORCH), nor are
the stdlib modules that were incidentally reachable through it; nothing names
any of them through this module. Only the tests move, and only where they reach
past the public surface.

Which is worth its own note. Two tests patched _read_sysfs_line_size on
tensor_class, and rebinding a re-exported name never affected the caller, whose
own globals are what it resolves against. One of them then passed on the real
detector returning the value it was asserting anyway, so the move is what
exposed it. They now patch the module that calls the function.

1452 lines becomes 1118 in the largest file, and the residency bookkeeping can
be read without the tensor surface layered over it.

Found in review.
