+++ python/utils/hostruntime/tensor_class.py
@@ -3,15 +3,30 @@
 # Copyright (C) 2025-2026 Advanced Micro Devices, Inc.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 #
-import bisect
+"""The host tensor: a shaped, typed view over a coherence-managed allocation.
+
+The allocation itself, its residency bookkeeping, and the torch bridge live in
+:mod:`buffer`, :mod:`coherence` and :mod:`torch_interop`. Each is re-exported
+here, so ``from .tensor_class import CPUBuffer`` and friends keep resolving.
+"""
+
 import math
 import operator
-import os
 from abc import ABC, abstractmethod
 from functools import cached_property
 
 import numpy as np
 import numpy.typing as npt
+
+from .buffer import CPUBuffer as CPUBuffer
+from .buffer import NpuBuffer as NpuBuffer
+from .buffer import _TensorBackedBuffer as _TensorBackedBuffer
+from .coherence import COHERENCE_GRANULE as COHERENCE_GRANULE
+from .coherence import _CoherenceMap as _CoherenceMap
+from .coherence import _detect_coherence_granule as _detect_coherence_granule
+from .coherence import _read_sysfs_line_size as _read_sysfs_line_size
+from .torch_interop import _array_to_torch as _array_to_torch
+from .torch_interop import _ml_dtype_to_torch_map as _ml_dtype_to_torch_map
 
 
 def _as_shape(shape):
@@ -38,355 +53,6 @@ def _as_shape(shape):
         if extent < 0:
             raise ValueError(f"shape extents must not be negative, got {extents}")
     return extents
-
-
-# Host<->device synchronization is not byte-granular: on a part that is not
-# cache-coherent the runtime maintains whole cache lines (it walks
-# ``sysconf(_SC_LEVEL1_DCACHE_LINESIZE)``-sized steps and the CPU acts on the
-# entire line containing an address), so a sync for one region unavoidably acts
-# on every byte sharing its first and last line. Sub-region views must therefore
-# not share a line, or one view's sync can write a neighbor's stale host copy
-# over data the device just produced. Lines, not pages: pages are the unit of
-# address translation and protection, lines are the unit of coherence.
-_SYSFS_CACHE_DIR = "/sys/devices/system/cpu/cpu0/cache"
-
-
-def _read_sysfs_line_size(cache_dir=_SYSFS_CACHE_DIR):
-    """Line size of the first level-1 *data* cache, or None.
-
-    Cache indices are not ordered by level or type, so select on both rather
-    than assuming index0 is L1d: on a machine that enumerates the instruction
-    cache first, that assumption reads the wrong line size.
-    """
-    for index in range(8):
-        base = f"{cache_dir}/index{index}"
-        try:
-            with open(f"{base}/level") as fp:
-                level = int(fp.read().strip())
-            with open(f"{base}/type") as fp:
-                kind = fp.read().strip()
-            if level != 1 or kind not in ("Data", "Unified"):
-                continue
-            with open(f"{base}/coherency_line_size") as fp:
-                return int(fp.read().strip())
-        except (OSError, ValueError):
-            continue
-    return None
-
-
-def _detect_coherence_granule(default=64):
-    """Largest plausible coherence granule for this host, never below ``default``.
-
-    ``sysconf`` first because it is the portable spelling and is what the
-    runtime's own flush loop uses; the sysfs scan is a fallback for platforms
-    that do not report it. Rounded up to a power of two so the value can be used
-    as an alignment, and floored so a missing or nonsensical report can only ever
-    make the check stricter, never weaker. Absent both sources (notably Windows,
-    where ``os.sysconf`` does not exist) the floor is what is used, which is the
-    line size of every architecture this runs on today.
-    """
-    reported = None
-    try:
-        reported = os.sysconf("SC_LEVEL1_DCACHE_LINESIZE")
-    except (ValueError, OSError, AttributeError):
-        pass
-    if not reported or reported < 0:
-        reported = _read_sysfs_line_size()
-    granule = max(reported or 0, default)
-    # Round up to a power of two: alignment arithmetic assumes it, and a cache
-    # line is one on every architecture, but nothing in the reporting path
-    # promises it.
-    return 1 << (granule - 1).bit_length()
-
-
-COHERENCE_GRANULE = _detect_coherence_granule()
-
-# Mapping from ml_dtypes (non-native numpy) types to their torch equivalents.
-# Native numpy dtypes (float32, int32, …) are handled directly by torch.from_numpy
-# and do not need an entry here.
-# Populated lazily at first use to avoid importing torch/ml_dtypes at module load.
-_ML_DTYPE_TO_TORCH: dict | None = None
-
-
-def _ml_dtype_to_torch_map():
-    global _ML_DTYPE_TO_TORCH
-    if _ML_DTYPE_TO_TORCH is None:
-        import ml_dtypes
-        import torch  # pyright: ignore[reportMissingImports]
-
-        _candidates = {
-            ml_dtypes.bfloat16: torch.bfloat16,
-        }
-        for attr in (
-            "float8_e4m3fn",
-            "float8_e5m2",
-            "float8_e4m3fnuz",
-            "float8_e5m2fnuz",
-        ):
-            ml_dt = getattr(ml_dtypes, attr, None)
-            torch_dt = getattr(torch, attr, None)
-            if ml_dt is not None and torch_dt is not None:
-                _candidates[ml_dt] = torch_dt
-        _ML_DTYPE_TO_TORCH = {
-            np.dtype(ml_dt): torch_dt for ml_dt, torch_dt in _candidates.items()
-        }
-    return _ML_DTYPE_TO_TORCH
-
-
-# Same-width unsigned integer dtype for the ND reinterpret-view trick.
-_UINT_VIEW_DTYPE = {
-    1: np.uint8,
-    2: np.uint16,
-    4: np.uint32,
-    8: np.uint64,
-}
-
-
-def _array_to_torch(array: np.ndarray):
-    """
-    Convert a numpy array to a torch tensor, zero-copy.
-
-    For native numpy dtypes (float32, float16, int32, …) torch.from_numpy is used directly
-    (fastest path for these types).
-
-    For ml_dtypes types (bfloat16, float8_*) that torch cannot consume via from_numpy:
-    reinterpret as a same-width unsigned integer numpy view, wrap with from_numpy,
-    then view as the target torch dtype.  This is guaranteed zero-copy for all ranks.
-
-    Raises:
-        ImportError: If torch is not installed.
-    """
-    # _ml_dtype_to_torch_map() imports torch (raising ImportError with a helpful message
-    # if absent) and returns the ml_dtype -> torch dtype mapping.
-    torch_dtype = _ml_dtype_to_torch_map().get(array.dtype)
-    import torch  # pyright: ignore[reportMissingImports]  # already imported by _ml_dtype_to_torch_map(); cached by Python
-
-    if torch_dtype is None:
-        # Native numpy dtype: torch.from_numpy handles it directly and fastest.
-        return torch.from_numpy(array)
-
-    # ml_dtype: reinterpret memory as a same-width uint, then view as the torch dtype.
-    uint_dtype = _UINT_VIEW_DTYPE[array.dtype.itemsize]
-    return torch.from_numpy(array.view(uint_dtype)).view(torch_dtype)
-
-
-class _CoherenceMap:
-    """Which byte ranges of one allocation the host has written, and which the
-    device has.
-
-    The state belongs to the memory, not to whichever tensor happens to name it.
-    Ranges are kept coalesced, so an arena whose windows are all in the same
-    state costs one entry, and a whole-buffer reconcile is one operation rather
-    than one per view.
-    """
-
-    HOST = "cpu"
-    DEVICE = "npu"
-
-    def __init__(self, nbytes, state, granule=None):
-        self._spans = [[0, nbytes, state]]
-        # Where each span ends, kept beside the spans purely so a lookup can
-        # bisect plain integers. Bisecting the spans themselves needs a key
-        # function, which is a Python call at every step of the search.
-        self._ends = [nbytes]
-        self._granule = COHERENCE_GRANULE if granule is None else granule
-        # The whole allocation in one state, or None when it is mixed. Kept as a
-        # field because it answers most queries outright, and the queries are on
-        # the path a dispatch takes for every argument.
-        self.uniform = state
-
-    def set(self, start, end, state):
-        """Record ``[start, end)`` as ``state``.
-
-        Only the spans the range actually touches are rewritten. Rebuilding the
-        whole list instead is quadratic over a buffer written a window at a
-        time, which is the case this exists to serve.
-        """
-        if start >= end:
-            return
-        spans = self._spans
-        # First span ending after start, and first span beginning at or after
-        # end: everything between them is what this write disturbs.
-        ends = self._ends
-        # ends[i] is where span i stops, so the first span this write disturbs
-        # is the first one stopping after start, and the last is the one holding
-        # end - 1.
-        first = bisect.bisect_right(ends, start)
-        last = bisect.bisect_left(ends, end) + 1
-        replacement = []
-        if first < last:
-            head = spans[first]
-            if head[0] < start:  # keep the part of the first span before us
-                replacement.append([head[0], start, head[2]])
-            replacement.append([start, end, state])
-            tail = spans[last - 1]
-            if tail[1] > end:  # and the part of the last one after us
-                replacement.append([end, tail[1], tail[2]])
-        else:
-            replacement.append([start, end, state])
-        # Merge with the neighbours on either side, so a buffer written window
-        # by window collapses back to one span once the windows meet.
-        while first > 0 and spans[first - 1][2] == replacement[0][2]:
-            replacement[0][0] = spans[first - 1][0]
-            first -= 1
-        while last < len(spans) and spans[last][2] == replacement[-1][2]:
-            replacement[-1][1] = spans[last][1]
-            last += 1
-        merged = [replacement[0]]
-        for span in replacement[1:]:
-            if merged[-1][2] == span[2]:
-                merged[-1][1] = span[1]
-            else:
-                merged.append(span)
-        spans[first:last] = merged
-        ends[first:last] = [span[1] for span in merged]
-        self._check_granule(merged, start, end, state)
-        self.uniform = spans[0][2] if len(spans) == 1 else None
-
-    def _check_granule(self, merged, start, end, state):
-        """No coherence granule may hold regions in two different states.
-
-        This is what makes the map mean anything: reconciling one range acts on
-        whole cache lines, so two states sharing a line cannot be maintained
-        independently and one of them will be written over the other.
-
-        Only the boundaries this write introduces are checked. Every other
-        boundary was checked when it was made and nothing here moves it, so the
-        property is carried forward rather than re-established, which keeps a
-        buffer written window by window from costing more with each window. The
-        ends of the allocation are not boundaries: nothing lies beyond them.
-        """
-        granule = self._granule
-        # Every edge of the rewritten block: where it meets its neighbours, and
-        # the boundaries inside it. At most a handful, whatever the buffer holds.
-        edges = {merged[0][0]}
-        edges.update(span[1] for span in merged)
-        end_of_buffer = self._spans[-1][1]
-        for edge in edges:
-            if edge and edge != end_of_buffer and edge % granule:
-                raise ValueError(
-                    f"recording {state!r} for bytes [{start}, {end}) would leave a "
-                    f"state boundary at byte {edge}, which is not on a {granule}-byte "
-                    f"coherence granule. Host and device are reconciled a cache line "
-                    f"at a time, so the regions either side of that boundary could "
-                    f"not be transferred independently."
-                )
-
-    def get(self, start, end):
-        """The state of ``[start, end)``, or None if it is not uniform."""
-        if self.uniform is not None:
-            return self.uniform
-        seen = {v for lo, hi, v in self._spans if lo < end and hi > start}
-        return seen.pop() if len(seen) == 1 else None
-
-    def any_host_written(self, start, end):
-        if self.uniform is not None:
-            return self.uniform == self.HOST
-        return any(
-            v == self.HOST for lo, hi, v in self._spans if lo < end and hi > start
-        )
-
-    def ranges(self, start, end, state):
-        """The sub-ranges of ``[start, end)`` currently in ``state``."""
-        return [
-            (max(lo, start), min(hi, end))
-            for lo, hi, value in self._spans
-            if value == state and lo < end and hi > start
-        ]
-
-
-class NpuBuffer(ABC):
-    """One allocation, and the coherence of the memory in it.
-
-    A buffer owns bytes and knows how to reconcile them between host and device.
-    It has no shape and no dtype: those are interpretations, and they belong to
-    the :class:`NpuTensor` views layered over it. Many tensors may share one
-    buffer, which is why the coherence state lives here -- it describes the
-    memory, so storing it per tensor would let two names for the same bytes
-    disagree.
-
-    Compare :class:`torch.Storage`, which draws the same line for the same
-    reason, and is where ``storage_offset`` comes from.
-    """
-
-    def __init__(self, nbytes, device, granule=None):
-        self.nbytes = nbytes
-        self.coherence = _CoherenceMap(nbytes, device, granule)
-
-    @property
-    @abstractmethod
-    def host_bytes(self) -> np.ndarray:
-        """The allocation as a flat ``uint8`` array the host can address.
-
-        Every tensor over this buffer takes its data from here, so it is part of
-        the contract rather than a convention each backend happens to follow.
-        """
-
-    @abstractmethod
-    def sync_to_device(self, offset, nbytes):
-        """Make the host's writes to ``[offset, offset+nbytes)`` visible to the device."""
-
-    @abstractmethod
-    def sync_from_device(self, offset, nbytes):
-        """Make the device's writes to ``[offset, offset+nbytes)`` visible to the host."""
-
-    def binding_handle(self, offset, nbytes):
-        """A handle a runtime can bind for this region, if the backend has one.
-
-        Returning None is a legitimate answer, not a stub: a design where the
-        host writes at offsets into a whole allocation and the kernel addresses
-        the layout itself never needs a per-region handle at all.
-        """
-        return None
-
-
-class _TensorBackedBuffer(NpuBuffer):
-    """Compatibility buffer for a backend that still owns its own storage.
-
-    Delegates reconciliation to the tensor's transfer methods, whole-extent,
-    which is what those backends did before buffers existed. It lets a backend
-    adopt the split when it is ready rather than being restructured by this
-    change, without opting out of per-region coherence.
-    """
-
-    def __init__(self, tensor):
-        super().__init__(
-            tensor.nbytes,
-            tensor._initial_device,
-            type(tensor)._resolve_coherence_granule(),
-        )
-        self._tensor = tensor
-
-    def sync_to_device(self, offset, nbytes):
-        self._tensor._sync_to_device()
-
-    def sync_from_device(self, offset, nbytes):
-        self._tensor._sync_from_device()
-
-    def binding_handle(self, offset, nbytes):
-        return getattr(self._tensor, "_bo", None)
-
-    @property
-    def host_bytes(self):
-        return self._tensor._data.reshape(-1).view(np.uint8)
-
-
-class CPUBuffer(NpuBuffer):
-    """A host allocation. Reconciliation is a no-op: there is only one agent."""
-
-    def __init__(self, nbytes, device, granule=None):
-        super().__init__(nbytes, device, granule)
-        self._host = np.zeros(nbytes, dtype=np.uint8)
-
-    @property
-    def host_bytes(self):
-        return self._host
-
-    def sync_to_device(self, offset, nbytes):
-        pass
-
-    def sync_from_device(self, offset, nbytes):
-        pass
 
 
 class _WriteBorrow:
